requirements.txt
jupyter
ipykernel
pandas
numpy
matplotlib
pillow
scikit-learn
torch
torchvision
mlflow
dvc
huggingface_hub
